test: Implement a default repository class to be extended.
