+++ alert-database/src/test/java/com/synopsys/integration/alert/database/api/mock/DefaultMockJPARepository.java
@@ -1,0 +1,140 @@
+package com.synopsys.integration.alert.database.api.mock;
+
+import java.util.List;
+import java.util.Optional;
+
+import org.springframework.data.domain.Example;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public class DefaultMockJPARepository<T, ID> implements JpaRepository<T,ID> {
+    // Empty implementation of a JPA repository to be extended by other mock repositories.
+    // The child classes that extend this can then override the methods they need to
+    // provide the correct testing functionality needed.
+    // To use the class correctly have a child class extend this class and also implement the JPA Repository
+    // interface to be tested.
+    //
+    // For example creating a mock for SystemStatusRepository
+    // class MockSystemStatusRepository extends DefaultMockJPARepository<SystemStatus, Long> implements SystemStatusRepository
+    @Override
+    public List<T> findAll() {
+        return null;
+    }
+
+    @Override
+    public List<T> findAll(Sort sort) {
+        return null;
+    }
+
+    @Override
+    public List<T> findAllById(Iterable<ID> ids) {
+        return null;
+    }
+
+    @Override
+    public <S extends T> List<S> saveAll(Iterable<S> entities) {
+        return null;
+    }
+
+    @Override
+    public void flush() {
+
+    }
+
+    @Override
+    public <S extends T> S saveAndFlush(S entity) {
+        return null;
+    }
+
+    @Override
+    public void deleteInBatch(Iterable<T> entities) {
+
+    }
+
+    @Override
+    public void deleteAllInBatch() {
+
+    }
+
+    @Override
+    public T getOne(ID id) {
+        return null;
+    }
+
+    @Override
+    public <S extends T> List<S> findAll(Example<S> example) {
+        return null;
+    }
+
+    @Override
+    public <S extends T> List<S> findAll(Example<S> example, Sort sort) {
+        return null;
+    }
+
+    @Override
+    public Page<T> findAll(Pageable pageable) {
+        return null;
+    }
+
+    @Override
+    public <S extends T> S save(S entity) {
+        return null;
+    }
+
+    @Override
+    public Optional<T> findById(ID id) {
+        return Optional.empty();
+    }
+
+    @Override
+    public boolean existsById(ID id) {
+        return false;
+    }
+
+    @Override
+    public long count() {
+        return 0;
+    }
+
+    @Override
+    public void deleteById(ID id) {
+
+    }
+
+    @Override
+    public void delete(T entity) {
+
+    }
+
+    @Override
+    public void deleteAll(Iterable<? extends T> entities) {
+
+    }
+
+    @Override
+    public void deleteAll() {
+
+    }
+
+    @Override
+    public <S extends T> Optional<S> findOne(Example<S> example) {
+        return Optional.empty();
+    }
+
+    @Override
+    public <S extends T> Page<S> findAll(Example<S> example, Pageable pageable) {
+        return null;
+    }
+
+    @Override
+    public <S extends T> long count(Example<S> example) {
+        return 0;
+    }
+
+    @Override
+    public <S extends T> boolean exists(Example<S> example) {
+        return false;
+    }
+}
